Some minor cleaning up.

+++ src/plugins/support/PMRSupport/src/pmrworkspace.cpp
@@ -620,7 +620,7 @@ void PmrWorkspace::refreshStatus()
                                                nullptr;
 
                 if (filePath) {
-                    CharPair statusChars = gitStatusChars(status->status);
+                    CharPair statusChars = gitStatusChars(uint(status->status));
 
                     if (statusChars.first != ' ')
                         ++mStagedCount;
@@ -910,7 +910,7 @@ void PmrWorkspace::stageFile(const QString &pPath, bool pStage)
                             memset(&indexEntry, '\0', sizeof(git_index_entry));
 
                             indexEntry.id = *git_tree_entry_id(headEntry);
-                            indexEntry.mode = git_tree_entry_filemode(headEntry);
+                            indexEntry.mode = uint32_t(git_tree_entry_filemode(headEntry));
                             indexEntry.path = relativePath;
 
                             git_index_add(index, &indexEntry);
